+++ packages/proxy/lib/http/request-middleware.ts
@@ -157,7 +157,9 @@ const MaybeEndRequestWithBufferedResponse: RequestMiddleware = function () {
 
   if (buffer) {
     this.debug('ending request with buffered response')
-    this.res.wantsInjection = this.config.experimentalSessionAndOrigin && buffer.isCrossOrigin ? 'fullCrossOrigin' : 'full'
+    // NOTE: Only inject fullCrossOrigin here if experimental is on and
+    // the super domain origins do not match in order to keep parity with cypress application reloads
+    this.res.wantsInjection = this.config.experimentalSessionAndOrigin && buffer.isCrossSuperDomainOrigin ? 'fullCrossOrigin' : 'full'
 
     return this.onResponse(buffer.response, buffer.stream)
   }