+++ packages/proxy/test/unit/http/request-middleware.spec.ts
@@ -135,7 +135,7 @@ describe('http/request-middleware', () => {
         },
         getAUTUrl: sinon.stub().returns('http://localhost:8080'),
         remoteStates: {
-          isPrimaryOrigin: sinon.stub().returns(false),
+          isPrimarySuperDomainOrigin: sinon.stub().returns(false),
         },
         req: {
           headers: {
@@ -159,7 +159,7 @@ describe('http/request-middleware', () => {
         },
         getAUTUrl: sinon.stub().returns('http://localhost:8080'),
         remoteStates: {
-          isPrimaryOrigin: sinon.stub().returns(false),
+          isPrimarySuperDomainOrigin: sinon.stub().returns(false),
         },
         resourceTypeAndCredentialManager: {
           get: sinon.stub().returns({}),
@@ -186,7 +186,7 @@ describe('http/request-middleware', () => {
         },
         getAUTUrl: sinon.stub().returns('http://localhost:8080'),
         remoteStates: {
-          isPrimaryOrigin: sinon.stub().returns(false),
+          isPrimarySuperDomainOrigin: sinon.stub().returns(false),
         },
         resourceTypeAndCredentialManager: {
           get: sinon.stub().returns({}),
@@ -212,7 +212,7 @@ describe('http/request-middleware', () => {
         },
         getAUTUrl: sinon.stub().returns('http://localhost:8080'),
         remoteStates: {
-          isPrimaryOrigin: sinon.stub().returns(false),
+          isPrimarySuperDomainOrigin: sinon.stub().returns(false),
         },
         resourceTypeAndCredentialManager: {
           get: sinon.stub().returns({
@@ -338,7 +338,7 @@ describe('http/request-middleware', () => {
       const ctx = await getContext()
 
       ctx.req.isAUTFrame = false
-      ctx.remoteStates.isPrimaryOrigin.returns(true),
+      ctx.remoteStates.isPrimarySuperDomainOrigin.returns(true),
 
       await testMiddleware([MaybeAttachCrossOriginCookies], ctx)
 
@@ -529,7 +529,7 @@ describe('http/request-middleware', () => {
         getAUTUrl: () => autUrl,
         getCookieJar: () => cookieJar,
         remoteStates: {
-          isPrimaryOrigin: sinon.stub().returns(false),
+          isPrimarySuperDomainOrigin: sinon.stub().returns(false),
         },
         config: { experimentalSessionAndOrigin: true },
         req: {
@@ -548,7 +548,7 @@ describe('http/request-middleware', () => {
 
     it('sets wantsInjection to full when a request is buffered', async () => {
       const buffers = new HttpBuffers()
-      const buffer = { url: 'https://www.cypress.io/', isCrossOrigin: false } as HttpBuffer
+      const buffer = { url: 'https://www.cypress.io/', isCrossSuperDomainOrigin: false } as HttpBuffer
 
       buffers.set(buffer)
 
@@ -568,7 +568,7 @@ describe('http/request-middleware', () => {
 
     it('sets wantsInjection to fullCrossOrigin when a cross origin request is buffered and experimentalSessionAndOrigin=true', async () => {
       const buffers = new HttpBuffers()
-      const buffer = { url: 'https://www.cypress.io/', isCrossOrigin: true } as HttpBuffer
+      const buffer = { url: 'https://www.cypress.io/', isCrossSuperDomainOrigin: true } as HttpBuffer
 
       buffers.set(buffer)
 
@@ -591,7 +591,7 @@ describe('http/request-middleware', () => {
 
     it('sets wantsInjection to full when a cross origin request is buffered and experimentalSessionAndOrigin=false', async () => {
       const buffers = new HttpBuffers()
-      const buffer = { url: 'https://www.cypress.io/', isCrossOrigin: true } as HttpBuffer
+      const buffer = { url: 'https://www.cypress.io/', isCrossSuperDomainOrigin: true } as HttpBuffer
 
       buffers.set(buffer)
 
@@ -614,7 +614,7 @@ describe('http/request-middleware', () => {
 
     it('wantsInjection is not set when the request is not buffered', async () => {
       const buffers = new HttpBuffers()
-      const buffer = { url: 'https://www.cypress.io/', isCrossOrigin: true } as HttpBuffer
+      const buffer = { url: 'https://www.cypress.io/', isCrossSuperDomainOrigin: true } as HttpBuffer
 
       buffers.set(buffer)